v4l2codecs: encoder: Implement zero-copy support
In this implementation we switch encoders to always run in memory type
V4L2_MEMORY_DMABUF. To alllow fallback, we allocated 4 frame using
V4L2_MEMORY_MMAP at setup time and detach the dmabuf from the driver. If
upstream provides dmabuf, we try importating that, and if that fails, we
make a copy and continue.