HACK: v4l2codecs: ensure that buffers are not reused
The encoder element may reuse buffers for importing dmabuf_fds. This causes the
driver to not use all V4L2 buffers but may result in the reuse of a single
buffer for all frames, which is OK, but may be surprising during debugging.

Avoid the surprise and ensure that all buffers are rotated by only using buffers
that are not initializes with a dmabuf, yet.