v4l2codecs: encoder: fix S_FMT on CAPTURE queue
The stateful encoder API ignores the width and height on the CAPTURE queue and
configures it only on the OUTPUT queue (for the raw buffers).

Don't set the width and height for the stateless API, too.

Always update the format to ensure that sizeimage is used for the CAPTURE
buffers.